Record length changed from integer(8)(is not supported by many compilers) to integer(4). GW DOS of states is added. GW and Spin-Pol tutorials has been successfully tested.

+++ src/src_gw/calcepsilon.f90
@@ -34,7 +34,7 @@ subroutine calcepsilon(iqp,COMM_LEVEL2)
       integer(4) :: i, isym, lspl, isym0
       integer(4) :: im, jm
       integer(4) :: dimtk
-      integer(8) :: Recl
+      integer(4) :: Recl
       integer:: level2rank,level2procs
